requirements.txt
Flask
pytesseract
Pillow
pdf2image
PyPDF2